+++ src/pages/Resume.tsx
@@ -1,3 +1,4 @@
+import ResumeButtons from "../components/ResumeButtons";
 import Skills from "../components/Skills";
 
 export default function Resume() {
@@ -5,6 +6,12 @@ export default function Resume() {
   return (
     <>
       <div className="bg-gray-800 text-white">
+        <h2 className="text-center">
+          Here&apos;s my resume and a list of my skills
+        </h2>
+
+        <ResumeButtons />
+
         <Skills />
       </div>
     </>
